Fix pyramid writing error (see commit description)
When building the image pyramid, we "read back" level N from the file we are writing to to make level N+1. In tifffile - "Default number of bytes to read or encode in one pass (256 MB)". When the image is smaller than the 256 MB and wasn't flushed to the file when we try to read it back, it errors out. So we enforce flushing from python buffer to file between each pyramid levels.